changed ex1 datatype

+++ src/ex1.c
@@ -3,8 +3,8 @@
 
 int main(int argc, char const *argv[])
 {
-long int num1 = 42;
-long int num2 = 48879;
+int num1 = 42;
+int num2 = 48879;
 char str[] = "AAABBB";
 
 printf(argv[1]);
